# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小知，你的校园助手",
        "title": "KnowEase API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/posts/{postid}/{userid}/save": {
            "post": {
                "description": "用户给帖子收藏，收藏数加一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-收藏"
                ],
                "summary": "用户收藏帖子",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户收藏成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "收藏记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/cancelsave": {
            "post": {
                "description": "用户取消收藏，收藏数减一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-收藏"
                ],
                "summary": "用户取消收藏帖子",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户取消收藏成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "取消收藏记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/getcounts": {
            "get": {
                "description": "获取帖子的点赞数，收藏数，浏览量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子"
                ],
                "summary": "获取帖子相关数值",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "相关数值",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}/like": {
            "post": {
                "description": "用户取消点赞，点赞数减一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-点赞"
                ],
                "summary": "用户取消点赞帖子",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户D",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户取消点赞成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "取消点赞记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/{commentid}/cancellike": {
            "post": {
                "description": "用户取消点赞评论，评论数减一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "用户取消点赞评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户取消点赞成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "取消点赞记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}/{commentid}/getcounts": {
            "get": {
                "description": "获取评论的点赞数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "获取评论点赞数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "相关数值",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}/{commentid}/like": {
            "post": {
                "description": "用户给评论点赞，评论数加一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "用户点赞评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户收藏成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "收藏记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/{commentid}/{replyid}/cancellike": {
            "post": {
                "description": "用户取消点赞回复，点赞数减一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "用户取消点赞回复",
                "parameters": [
                    {
                        "type": "string",
                        "description": "回复ID",
                        "name": "replyid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户取消点赞成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "取消点赞记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}/{commentid}/{replyid}/getcounts": {
            "get": {
                "description": "获取回复的点赞数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "获取回复点赞数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "回复ID",
                        "name": "replyid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "相关数值",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}/{commentid}/{replyid}/like": {
            "post": {
                "description": "用户给回复点赞，评论数加一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "用户点赞回复",
                "parameters": [
                    {
                        "type": "string",
                        "description": "回复ID",
                        "name": "replyid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户点赞成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "点赞记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/userpage/likecount": {
            "get": {
                "description": "用户给帖子点赞，点赞数加一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人主页"
                ],
                "summary": "获取用户的总获赞数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询用户获赞数成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "查询用户获赞数失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/userpage/likerecord": {
            "get": {
                "description": "在个人主页获取用户一个月（31天）内的点赞记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人主页-记录查询"
                ],
                "summary": "获取用户历史一个月内的历史点赞记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功信息以及历史点赞帖子",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "获取点赞记录失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}